рефакторинг

+++ application/models/upload-file_model.php
@@ -11,14 +11,15 @@ class UploadFileModel extends \core\Model
 
         // поиск других загрузок изображений этого профиля и установка нового имени файла
         $dwlDirPath = dirname(__DIR__, 1)."\data\\temp\\"; // папка, куда перемещается изображение из $_POST
-        $dwlfiles = glob("$dwlDirPath$email*");
+        $dwlfiles = glob($dwlDirPath.$email.'*'); // поиск файлов по шаблону
+        $number = false;
         if(count($dwlfiles) == 0){
             $filename = $email.'.1.'.$ext;
         }
         else{
-            $dwlFile = $dwlfiles[count($dwlfiles)-1];
-            $dwlfiles = explode('.', $dwlFile);
-            $number = intval($dwlfiles[count($dwlfiles)-2]);
+            $dwlFile = $dwlfiles[count($dwlfiles)-1];           // последний файл
+            $dwlfiles = explode('.', $dwlFile);                 // имя разбивается по точкам в массив
+            $number = intval($dwlfiles[count($dwlfiles)-2]);    // предпоследний индекс файла
             $filename = $email.'.'.++$number.'.'.$ext;
         }
 
@@ -27,10 +28,12 @@ class UploadFileModel extends \core\Model
         $toPath =  "$dwlDirPath$filename"; // куда перемещается
         echo move_uploaded_file($fromPath, $toPath) ? $filename : '';
         // удаление предыдущих вариантов изображения
-        for($i=1; $i<$number; $i++){
-            $filename = "$dwlDirPath$email.$i.$ext";
-            if(file_exists($filename)){ 
-                unlink( "$dwlDirPath$email.$i.$ext");
+        if($number){
+            for($i=1; $i<$number; $i++){
+                $filename = "$dwlDirPath$email.$i.$ext";
+                if(file_exists($filename)){ 
+                    unlink( "$dwlDirPath$email.$i.$ext");
+                }
             }
         }
     }
